v. 085
### Project: zwasnki_bot

**Description:**
This Telegram bot, developed using the `python-telegram-bot` library, initiates a conversation with the user and provides a custom keyboard interface with the following options:

- **IMEI Services**
- **Check Your Account Balance**
- **Visit Server for More**

Upon starting a chat with the bot and sending the `/start` command, the bot prompts the user to select an action from the provided options. Depending on the user's selection, the bot responds accordingly, providing the requested service or information.

**Features:**
- Custom keyboard interface for easy navigation
- Handles user input to execute corresponding actions
- Implements secure communication via the Telegram platform
- Developed using the `python-telegram-bot` library for seamless integration with Telegram's API

**Usage:**
1. Start a chat with the bot by sending the `/start` command.
2. Select an action from the provided options: IMEI Services, Check Your Account Balance, or Visit Server for More.
3. Follow the prompts to access the desired service or information.

**How to Run:**
1. Clone the repository from [zwanski2019/zwasnki_bot](https://github.com/zwanski2019/zwasnki_bot).
2. Replace `'YOUR_BOT_TOKEN'` with your actual bot token obtained from BotFather.
3. Run the main Python script `zwanski_bot.py` to start the bot.

**Dependencies:**
- `python-telegram-bot`: The Python wrapper for the Telegram Bot API used for bot development.

**Contributing:**
Contributions are welcome! Feel free to fork the repository, make improvements, and submit pull requests.

**License:**
This project is licensed under the [MIT License](https://opensource.org/licenses/MIT).

**Contact:**
For any inquiries or feedback, please contact Mohammed Ibrahim A (Zwanski) at [your@email.com]() https://linktr.ee/zwanski